refactor common code into base classes
