+++ trajoptlib/include/trajopt/SwerveTrajectoryGenerator.hpp
@@ -10,6 +10,7 @@
 
 #include <sleipnir/optimization/OptimizationProblem.hpp>
 
+#include "trajopt/TrajectoryGenerator.hpp"
 #include "trajopt/path/SwervePathBuilder.hpp"
 #include "trajopt/solution/SwerveSolution.hpp"
 #include "trajopt/util/SymbolExports.hpp"
@@ -21,7 +22,7 @@ namespace trajopt {
  * This trajectory generator class contains functions to generate
  * time-optimal trajectories for several drivetrain types.
  */
-class TRAJOPT_DLLEXPORT SwerveTrajectoryGenerator {
+class TRAJOPT_DLLEXPORT SwerveTrajectoryGenerator : TrajectoryGenerator {
  public:
   /**
    * Construct a new swerve trajectory optimization problem.
